Converting node locs to doubles

+++ Source/mesquite/lib/TreeDrawing.java
@@ -331,7 +331,7 @@ public abstract class TreeDrawing  {
 	public abstract int getEdgeWidth();
 	
 	/** project point x,y onto the line between N's lineBase to lineTip. */
-	public Point2D projectionOnLine(int N, int x, int y) {
+	public Point2D projectionOnLine(int N, double x, double y) {
 		double newX = x;
 		double newY = y;
 		try {
